requirements.txt
matplotlib
art
Pillow
subword-nmt